separate parser tests from parser service
